% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphene_spp_dispersion.R
\name{beta}
\alias{beta}
\title{Coupling Term}
\usage{
beta(q, omega, d = 3.5e-10, eps_inf, omega_TO, omega_LO)
}
\arguments{
\item{q}{wavevector [m^-1]}

\item{omega}{angular frequency [s^-1]}

\item{d}{distance between graphene and substrate [m]}

\item{eps_inf}{permittivity at infinity / high frequency dielectric constant}
}
\value{
returns the coupling term beta
}
\description{
The coupling term between uncoupled conductive sheet and phonons
}
\references{
[1] Wang, S Q. & Mahan G. D. Phys. Rev. B (1972) 6 4517

[2] Hwang, E. H., Sensarma, R. & Das Sarma, S. Phys. Rev. B (2010), 82, 195406
}

