% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphene_spp_dispersion.R
\name{eps_total}
\alias{eps_total}
\title{Total dielectric function}
\usage{
eps_total(q, omega, eps_inf, E_f, tau)
}
\arguments{
\item{q}{wavevector [m^-1]}

\item{omega}{angular frequency [s^-1]}

\item{eps_inf}{permittivity at infinity / high frequency dielectric constant}

\item{E_f}{Fermi energy [J]}

\item{tau}{relaxation time of electrons [s]}

\item{d}{distance between graphene and substrate [m]}
}
\value{
returns the total dielectirc function
#'
}
\description{
The total dielectric function within the random phase approximation
}
\references{
[1] Yan, H. et al. Nat. Photonics (2013), 7, 394

[2] Hwang, E. H., Sensarma, R. & Das Sarma, S. Phys. Rev. B (2010), 82, 195406
}

