\name{grapheneSPP-package}
\alias{grapheneSPP-package}
\alias{grapheneSPP}
\docType{package}
\title{
Graphene SPP Dispersion
}
\description{
A package that calculates the graphene loss function using the Drude model, for the plotting of the surface plasmon dispersion relation.
}
\details{
\tabular{ll}{
Package: \tab grapheneSPP\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-07\cr
License: \tab MIT + file LICENSE\cr
}

}
\author{
Tom Constant

Maintainer: <t.j.constant[at]gmail.com>

}
\references{
Luxmoore, I. J. et al. ACS Photonics (2014) 1 (11), 1151-1155

Low, T. & Avouris, P. ACS Nano (2014), 8, 1086

Gan, C. H. Appl. Phys. Lett. (2012) 101 111609

Yan, H. et al. Nat. Photonics (2013), 7, 394

Wang, S Q. & Mahan G. D. Phys. Rev. B (1972) 6 4517

Hwang, E. H., Sensarma, R. & Das Sarma, S. Phys. Rev. B (2010), 82, 195406
}

\keyword{ package }
