% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phonon_coupling.R
\name{phonon_coupling}
\alias{phonon_coupling}
\title{Phonon Coupling Position and Strength Calculation}
\usage{
phonon_coupling(q, omega, d = 3.5e-10, omega_TO, omega_LO,
  eps_graphene = 3.9, eps_substrate = 2.4)
}
\arguments{
\item{q}{wavevector [m^-1]}

\item{omega}{angular frequency [s^-1]}

\item{d}{distance between graphene and substrate [m]}

\item{omega_TO}{vector of frequencies of transverse optical phonons [cm^-1]}

\item{omega_LO}{vector of frequencies of longitudinal optical phonons [cm^-1]}

\item{eps_graphene}{static permittivity}

\item{eps_substrate}{permittivity at infinity / high frequency dielectric constant (also refered to elsewhere as eps_inf)}
}
\description{
Function to return surface phonon frequencies and coupling coupling coefficents for dispersion
}
\note{
Currently only 3 phonons supported
}

