% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphene_spp_dispersion.R
\name{pi_drude}
\alias{pi_drude}
\title{Simplifed Drude polarizability of uncoupled sheet}
\usage{
pi_drude(q, omega, eps_inf, E_f, tau)
}
\arguments{
\item{q}{wavevector [m^-1]}

\item{omega}{angular frequency [s^-1]}

\item{eps_inf}{permittivity at infinity / high frequency dielectric constant}

\item{E_f}{Fermi energy [J]}

\item{tau}{relaxation time of electrons [s]}

\item{d}{distance between graphene and substrate [m]}
}
\value{
returns the total dielectirc function
}
\description{
The polarizability of an uncoupled sheet described by the Drude model
}
\details{
Note the conductivity of graphene contains two contributions for interband and intraband (Drude) transistions.
This model assumes h_bar*omega < 2E_f, and so interband conductivity may be neglected.
}
\references{
[1] Low, T. & Avouris, P. ACS Nano (2014), 8, 1086

[2] Gan, C. H. Appl. Phys. Lett. (2012) 101 111609
}

