% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphene_spp_dispersion.R
\name{v_f}
\alias{v_f}
\title{2D Fourier transform of the Coulomb potential}
\usage{
v_f(q, eps_inf)
}
\arguments{
\item{q}{wavevector [m^-1]}

\item{eps_inf}{permittivity at infinity / high frequency dielectric constant}
}
\value{
2D Fourier transform of the Coulomb potential
}
\description{
2D Fourier transform of the Coulomb potential
}

