% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/line_extraction.R
\name{dispersion_extraction}
\alias{dispersion_extraction}
\title{Dispersion Extraction}
\usage{
dispersion_extraction(q, omega, loss_matrix)
}
\arguments{
\item{q}{wavevector [m^-1]}

\item{omega}{angular frequency [s^-1]}

\item{loss_matrix}{the log of the loss matrix}
}
\description{
Function to return sdispersion as points
}
\note{
Currently only 3 phonons supported
}
\examples{
omega<-seq(0,60,,500)*2*pi*1e12
q<-seq(0,2*pi/550e-9,,500)
plot_it<-function(Ef){
  loss_matrix<-log(loss_function(q,omega,eps_inf = 2.4,E_f = Ef,tau=1.5e-13))
  result<-dispersion_extraction(q,omega,loss_matrix)
  ggplot(result,aes(q,omega,color=factor(branch)))+geom_line(size=2)+theme_bw()+ylim(range(omega))
}
library(manipulate)
manipulate(plot_it(E_f),E_f=slider(min = 0.000001,max = 1))
}

