% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/line_extraction.R
\name{dispersion_extraction}
\alias{dispersion_extraction}
\title{Dispersion Extraction}
\usage{
dispersion_extraction(q, omega, loss_matrix)
}
\arguments{
\item{q}{wavevector [m^-1]}

\item{omega}{angular frequency [s^-1]}

\item{loss_matrix}{the log of the loss matrix}
}
\description{
Function to return sdispersion as points
}
\note{
Currently only 3 phonons supported
}

